+++ tests/test_layout_objects.py
@@ -165,8 +165,8 @@ def test_remove_labels():
     [
         ("text_input", "text_input.html"),
         ("text_area", "text_area.html"),
-        # ("checkboxes", "checkboxes.html"),
-        # ("radio", "radio.html"),
+        ("checkboxes", "checkboxes.html"),
+        ("radio", "radio.html"),
         ("single_checkbox", "single_checkbox.html"),
         ("select_input", "test_select.html"),
     ],